requirements.txt
wheel
scipy
websockets
asyncio
matplotlib
mysql-connector-python-rf
teamcity-messages